[lexical-code-core][lexical-code-shiki][lexical-code-prism] Feature: Outdent space-indented code lines
Shift+Tab in a code block currently only removes a leading TabNode, so
prettier-formatted code (which uses spaces for indentation) cannot be
outdented by the keyboard shortcut. This PR adds an optional tabSize
configuration; when set, the OUTDENT_CONTENT_COMMAND also strips up to
that many leading spaces from a code line.

When tabSize is undefined (the default), the original TabNode-only
outdent behavior is preserved.

Per line in the selection, with tabSize set:

- TabNode at the start: remove it (existing behavior takes
  precedence).
- CodeHighlightNode starting with spaces: strip
  min(tabSize, leadingSpaces). Best-effort, matching VS Code.
- Otherwise: no-op.

Selection is preserved relative to line content. When stripping K
characters, anchor/focus offsets pointing into the affected node
shift left by K (clamped to 0), so the cursor stays on the same
character rather than snapping to the start of the line.

Tab and INSERT_TAB_COMMAND continue to insert a TabNode regardless of
tabSize.

The keyboard / move handlers (Tab / Shift+Tab, alt+arrow line shifts,
Home/End) were duplicated verbatim between CodeHighlighterShiki and
CodeHighlighterPrism. This change moves them to a new
@lexical/code-core CodeIndentation module exposed in two ways:

- registerCodeIndentation(editor, tabSize?): a low-level register
  function. Used internally by the legacy registerCodeHighlighting
  wrapper, and exported standalone for callers who want indent
  without a highlighter.
- CodeIndentExtension: an Extension wrapping registerCodeIndentation
  with signal-based config (disabled, tabSize). Both
  CodeShikiExtension and CodePrismExtension declare it as a
  dependency, so Extension users configure tabSize on
  CodeIndentExtension.

To preserve backwards compatibility, the legacy registerCodeHighlighting
function keeps its existing signature and continues to register both
the highlighter transforms and the indent / Tab handlers as it always
did. The Extension flow uses a separate internal helper
(registerHighlightingOnly) for the highlighter side and depends on
CodeIndentExtension for the indent side, so the legacy export and the
Extensions are wired independently.

The space-stripping helper $outdentLeadingSpaces lives in
@lexical/code-core too, called from $handleMultilineIndent.

Both extensions' test suites use a shared $runOutdentScenario helper
and OUTDENT_SCENARIOS data table from
@lexical/code-core/__tests__/outdentTestUtils, driven through
test.for. Cases cover tabSize=2/4 with full and partial leading
spaces, the TabNode-precedence behavior, the tabSize=undefined
backward-compat path, cursor preservation, and rejection of
non-positive / non-integer tabSize.

Closes #8410
