+++ packages/lexical-code-core/flow/LexicalCodeCore.js.flow
@@ -73,6 +73,12 @@ declare export function $getLastCodeNodeOfLine(
   anchor: CodeHighlightNode | TabNode | LineBreakNode,
 ): CodeHighlightNode | TabNode | LineBreakNode;
 
+declare export function $outdentLeadingSpaces(
+  node: CodeHighlightNode,
+  tabSize: number,
+  selection: RangeSelection,
+): boolean;
+
 /**
  * CodeNode
  */
@@ -105,3 +111,15 @@ declare export class CodeNode extends ElementNode {
 }
 
 declare export var CodeExtension: LexicalExtension<ExtensionConfigBase, "@lexical/code", unknown, unknown>;
+
+export type CodeIndentConfig = {
+  disabled: boolean,
+  tabSize: number | void,
+};
+
+declare export function registerCodeIndentation(
+  editor: LexicalEditor,
+  tabSize?: number,
+): () => void;
+
+declare export var CodeIndentExtension: LexicalExtension<CodeIndentConfig, "@lexical/code-indent", unknown, unknown>;
