+++ packages/lexical-code-prism/src/CodeHighlighterPrism.ts
@@ -6,66 +6,33 @@
  *
  */
 
-import type {
-  BaseSelection,
-  LexicalCommand,
-  LexicalEditor,
-  LexicalNode,
-  LineBreakNode,
-  NodeKey,
-  RangeSelection,
-} from 'lexical';
+import type {LexicalEditor, LexicalNode, NodeKey} from 'lexical';
 
 import {
-  $getCodeLineDirection,
-  $getEndOfCodeInLine,
-  $getFirstCodeNodeOfLine,
-  $getLastCodeNodeOfLine,
-  $getStartOfCodeInLine,
   $isCodeHighlightNode,
   $isCodeNode,
   CodeExtension,
   CodeHighlightNode,
+  CodeIndentExtension,
   CodeNode,
   DEFAULT_CODE_LANGUAGE,
+  registerCodeIndentation,
 } from '@lexical/code-core';
 import {effect, namedSignals} from '@lexical/extension';
 import {
-  $createLineBreakNode,
-  $createPoint,
-  $createTabNode,
   $createTextNode,
-  $getCaretRange,
-  $getCaretRangeInDirection,
   $getNodeByKey,
-  $getRoot,
   $getSelection,
-  $getSiblingCaret,
-  $getTextPointCaret,
-  $insertNodes,
   $isLineBreakNode,
   $isRangeSelection,
   $isTabNode,
   $isTextNode,
-  $normalizeCaret,
   $onUpdate,
-  $setSelectionFromCaretRange,
-  COMMAND_PRIORITY_LOW,
   defineExtension,
-  INDENT_CONTENT_COMMAND,
-  INSERT_TAB_COMMAND,
-  KEY_ARROW_DOWN_COMMAND,
-  KEY_ARROW_UP_COMMAND,
-  KEY_TAB_COMMAND,
   mergeRegister,
-  MOVE_TO_END,
-  MOVE_TO_START,
-  OUTDENT_CONTENT_COMMAND,
   safeCast,
-  TabNode,
   TextNode,
 } from 'lexical';
-import invariant from 'shared/invariant';
 
 import {
   $getHighlightNodes,
@@ -334,416 +301,6 @@ function isEqual(nodeA: LexicalNode, nodeB: LexicalNode): boolean {
   );
 }
 
-/**
- * Returns a boolean.
- * Check that the selection span is within a single CodeNode.
- * This is used to guard against executing handlers that can only be
- * applied in a single CodeNode context
- */
-function $isSelectionInCode(selection: null | BaseSelection): boolean {
-  if (!$isRangeSelection(selection)) {
-    return false;
-  }
-  const anchorNode = selection.anchor.getNode();
-  const maybeAnchorCodeNode = $isCodeNode(anchorNode)
-    ? anchorNode
-    : anchorNode.getParent();
-  const focusNode = selection.focus.getNode();
-  const maybeFocusCodeNode = $isCodeNode(focusNode)
-    ? focusNode
-    : focusNode.getParent();
-
-  return (
-    $isCodeNode(maybeAnchorCodeNode) &&
-    maybeAnchorCodeNode.is(maybeFocusCodeNode)
-  );
-}
-
-/**
- * Returns an Array of code lines
- * Take the sequence of LineBreakNode | TabNode | CodeHighlightNode forming
- * the selection and split it by LineBreakNode.
- * If the selection ends at the start of the last line, it is considered empty.
- * Empty lines are discarded.
- */
-function $getCodeLines(
-  selection: RangeSelection,
-): Array<Array<CodeHighlightNode | TabNode>> {
-  const nodes = selection.getNodes();
-  const lines: Array<Array<CodeHighlightNode | TabNode>> = [];
-  if (nodes.length === 1 && $isCodeNode(nodes[0])) {
-    return lines;
-  }
-  let lastLine: Array<CodeHighlightNode | TabNode> = [];
-  for (let i = 0; i < nodes.length; i++) {
-    const node = nodes[i];
-    invariant(
-      $isCodeHighlightNode(node) || $isTabNode(node) || $isLineBreakNode(node),
-      'Expected selection to be inside CodeBlock and consisting of CodeHighlightNode, TabNode and LineBreakNode',
-    );
-    if ($isLineBreakNode(node)) {
-      if (lastLine.length > 0) {
-        lines.push(lastLine);
-        lastLine = [];
-      }
-    } else {
-      lastLine.push(node);
-    }
-  }
-  if (lastLine.length > 0) {
-    const selectionEnd = selection.isBackward()
-      ? selection.anchor
-      : selection.focus;
-
-    // Discard the last line if the selection ends exactly at the
-    // start of the line (no real selection)
-    const lastPoint = $createPoint(lastLine[0].getKey(), 0, 'text');
-    if (!selectionEnd.is(lastPoint)) {
-      lines.push(lastLine);
-    }
-  }
-
-  return lines;
-}
-
-function $handleTab(shiftKey: boolean): null | LexicalCommand<void> {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection) || !$isSelectionInCode(selection)) {
-    return null;
-  }
-  const indentOrOutdent = !shiftKey
-    ? INDENT_CONTENT_COMMAND
-    : OUTDENT_CONTENT_COMMAND;
-  const tabOrOutdent = !shiftKey ? INSERT_TAB_COMMAND : OUTDENT_CONTENT_COMMAND;
-
-  const anchor = selection.anchor;
-  const focus = selection.focus;
-
-  // 1. early decision when there is no real selection
-  if (anchor.is(focus)) {
-    return tabOrOutdent;
-  }
-
-  // 2. If only empty lines or multiple non-empty lines are selected: indent/outdent
-  const codeLines = $getCodeLines(selection);
-  if (codeLines.length !== 1) {
-    return indentOrOutdent;
-  }
-
-  const codeLine: Array<CodeHighlightNode | TabNode> = codeLines[0];
-  const codeLineLength = codeLine.length;
-
-  invariant(
-    codeLineLength !== 0,
-    '$getCodeLines only extracts non-empty lines',
-  );
-
-  // Take into account the direction of the selection
-  let selectionFirst;
-  let selectionLast;
-  if (selection.isBackward()) {
-    selectionFirst = focus;
-    selectionLast = anchor;
-  } else {
-    selectionFirst = anchor;
-    selectionLast = focus;
-  }
-
-  // find boundary elements of the line
-  // since codeLine only contains TabNode | CodeHighlightNode
-  // the result of these functions should is of Type TabNode | CodeHighlightNode
-  const firstOfLine = $getFirstCodeNodeOfLine(codeLine[0]);
-  const lastOfLine = $getLastCodeNodeOfLine(codeLine[0]);
-
-  const anchorOfLine = $createPoint(firstOfLine.getKey(), 0, 'text');
-  const focusOfLine = $createPoint(
-    lastOfLine.getKey(),
-    lastOfLine.getTextContentSize(),
-    'text',
-  );
-
-  // 3. multiline because selection started strictly before the line
-  if (selectionFirst.isBefore(anchorOfLine)) {
-    return indentOrOutdent;
-  }
-
-  // 4. multiline because the selection stops strictly after the line
-  if (focusOfLine.isBefore(selectionLast)) {
-    return indentOrOutdent;
-  }
-
-  // The selection if within the line.
-  // 4. If it does not touch both borders, it needs a tab
-  if (
-    anchorOfLine.isBefore(selectionFirst) ||
-    selectionLast.isBefore(focusOfLine)
-  ) {
-    return tabOrOutdent;
-  }
-
-  // 5. Selection is matching a full line on non-empty code
-  return indentOrOutdent;
-}
-
-function $handleMultilineIndent(type: LexicalCommand<void>): boolean {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection) || !$isSelectionInCode(selection)) {
-    return false;
-  }
-
-  const codeLines = $getCodeLines(selection);
-  const codeLinesLength = codeLines.length;
-
-  // Special Indent case
-  // Selection is collapsed at the beginning of a line
-  if (codeLinesLength === 0 && selection.isCollapsed()) {
-    if (type === INDENT_CONTENT_COMMAND) {
-      selection.insertNodes([$createTabNode()]);
-    }
-    return true;
-  }
-
-  // Special Indent case
-  // Selection is matching only one LineBreak
-  if (
-    codeLinesLength === 0 &&
-    type === INDENT_CONTENT_COMMAND &&
-    selection.getTextContent() === '\n'
-  ) {
-    const tabNode = $createTabNode();
-    const lineBreakNode = $createLineBreakNode();
-    const direction = selection.isBackward() ? 'previous' : 'next';
-    selection.insertNodes([tabNode, lineBreakNode]);
-    $setSelectionFromCaretRange(
-      $getCaretRangeInDirection(
-        $getCaretRange(
-          $getTextPointCaret(tabNode, 'next', 0),
-          $normalizeCaret($getSiblingCaret(lineBreakNode, 'next')),
-        ),
-        direction,
-      ),
-    );
-
-    return true;
-  }
-
-  // Indent Non Empty Lines
-  for (let i = 0; i < codeLinesLength; i++) {
-    const line = codeLines[i];
-    // a line here is never empty
-    if (line.length > 0) {
-      let firstOfLine: null | CodeHighlightNode | TabNode | LineBreakNode =
-        line[0];
-
-      // make sure to consider the first node on the first line
-      // because the line might not be fully selected
-      if (i === 0) {
-        firstOfLine = $getFirstCodeNodeOfLine(firstOfLine);
-      }
-
-      if (type === INDENT_CONTENT_COMMAND) {
-        const tabNode = $createTabNode();
-        firstOfLine.insertBefore(tabNode);
-        // First real code line may need selection adjustment
-        // when firstOfLine is at the selection boundary
-        if (i === 0) {
-          const anchorKey = selection.isBackward() ? 'focus' : 'anchor';
-          const anchorLine = $createPoint(firstOfLine.getKey(), 0, 'text');
-
-          if (selection[anchorKey].is(anchorLine)) {
-            selection[anchorKey].set(tabNode.getKey(), 0, 'text');
-          }
-        }
-      } else if ($isTabNode(firstOfLine)) {
-        firstOfLine.remove();
-      }
-    }
-  }
-  return true;
-}
-
-function $handleShiftLines(
-  type: LexicalCommand<KeyboardEvent>,
-  event: KeyboardEvent,
-): boolean {
-  // We only care about the alt+arrow keys
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) {
-    return false;
-  }
-
-  // I'm not quite sure why, but it seems like calling anchor.getNode() collapses the selection here
-  // So first, get the anchor and the focus, then get their nodes
-  const {anchor, focus} = selection;
-  const anchorOffset = anchor.offset;
-  const focusOffset = focus.offset;
-  const anchorNode = anchor.getNode();
-  const focusNode = focus.getNode();
-  const arrowIsUp = type === KEY_ARROW_UP_COMMAND;
-
-  // Ensure the selection is within the codeblock
-  if (
-    !$isSelectionInCode(selection) ||
-    !($isCodeHighlightNode(anchorNode) || $isTabNode(anchorNode)) ||
-    !($isCodeHighlightNode(focusNode) || $isTabNode(focusNode))
-  ) {
-    return false;
-  }
-  if (!event.altKey) {
-    // Handle moving selection out of the code block, given there are no
-    // siblings that can natively take the selection.
-    if (selection.isCollapsed()) {
-      const codeNode = anchorNode.getParentOrThrow();
-      if (
-        arrowIsUp &&
-        anchorOffset === 0 &&
-        anchorNode.getPreviousSibling() === null
-      ) {
-        const codeNodeSibling = codeNode.getPreviousSibling();
-        if (codeNodeSibling === null) {
-          codeNode.selectPrevious();
-          event.preventDefault();
-          return true;
-        }
-      } else if (
-        !arrowIsUp &&
-        anchorOffset === anchorNode.getTextContentSize() &&
-        anchorNode.getNextSibling() === null
-      ) {
-        const codeNodeSibling = codeNode.getNextSibling();
-        if (codeNodeSibling === null) {
-          codeNode.selectNext();
-          event.preventDefault();
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  let start;
-  let end;
-  if (anchorNode.isBefore(focusNode)) {
-    start = $getFirstCodeNodeOfLine(anchorNode);
-    end = $getLastCodeNodeOfLine(focusNode);
-  } else {
-    start = $getFirstCodeNodeOfLine(focusNode);
-    end = $getLastCodeNodeOfLine(anchorNode);
-  }
-  if (start == null || end == null) {
-    return false;
-  }
-
-  const range = start.getNodesBetween(end);
-  for (let i = 0; i < range.length; i++) {
-    const node = range[i];
-    if (
-      !$isCodeHighlightNode(node) &&
-      !$isTabNode(node) &&
-      !$isLineBreakNode(node)
-    ) {
-      return false;
-    }
-  }
-
-  // After this point, we know the selection is within the codeblock. We may not be able to
-  // actually move the lines around, but we want to return true either way to prevent
-  // the event's default behavior
-  event.preventDefault();
-  event.stopPropagation(); // required to stop cursor movement under Firefox
-
-  const linebreak = arrowIsUp
-    ? start.getPreviousSibling()
-    : end.getNextSibling();
-  if (!$isLineBreakNode(linebreak)) {
-    return true;
-  }
-  const sibling = arrowIsUp
-    ? linebreak.getPreviousSibling()
-    : linebreak.getNextSibling();
-  if (sibling == null) {
-    return true;
-  }
-
-  const maybeInsertionPoint =
-    $isCodeHighlightNode(sibling) ||
-    $isTabNode(sibling) ||
-    $isLineBreakNode(sibling)
-      ? arrowIsUp
-        ? $getFirstCodeNodeOfLine(sibling)
-        : $getLastCodeNodeOfLine(sibling)
-      : null;
-  let insertionPoint =
-    maybeInsertionPoint != null ? maybeInsertionPoint : sibling;
-  linebreak.remove();
-  range.forEach(node => node.remove());
-  if (type === KEY_ARROW_UP_COMMAND) {
-    range.forEach(node => insertionPoint.insertBefore(node));
-    insertionPoint.insertBefore(linebreak);
-  } else {
-    insertionPoint.insertAfter(linebreak);
-    insertionPoint = linebreak;
-    range.forEach(node => {
-      insertionPoint.insertAfter(node);
-      insertionPoint = node;
-    });
-  }
-
-  selection.setTextNodeRange(anchorNode, anchorOffset, focusNode, focusOffset);
-
-  return true;
-}
-
-function $handleMoveTo(
-  type: LexicalCommand<KeyboardEvent>,
-  event: KeyboardEvent,
-): boolean {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) {
-    return false;
-  }
-
-  const {anchor, focus} = selection;
-  const anchorNode = anchor.getNode();
-  const focusNode = focus.getNode();
-  const isMoveToStart = type === MOVE_TO_START;
-
-  // Ensure the selection is within the codeblock
-  if (
-    !$isSelectionInCode(selection) ||
-    !($isCodeHighlightNode(anchorNode) || $isTabNode(anchorNode)) ||
-    !($isCodeHighlightNode(focusNode) || $isTabNode(focusNode))
-  ) {
-    return false;
-  }
-
-  const focusLineNode = focusNode as CodeHighlightNode | TabNode;
-  const direction = $getCodeLineDirection(focusLineNode);
-  const moveToStart = direction === 'rtl' ? !isMoveToStart : isMoveToStart;
-
-  if (moveToStart) {
-    const start = $getStartOfCodeInLine(focusLineNode, focus.offset);
-    if (start !== null) {
-      const {node, offset} = start;
-      if ($isLineBreakNode(node)) {
-        node.selectNext(0, 0);
-      } else {
-        selection.setTextNodeRange(node, offset, node, offset);
-      }
-    } else {
-      focusLineNode.getParentOrThrow().selectStart();
-    }
-  } else {
-    const node = $getEndOfCodeInLine(focusLineNode);
-    node.select();
-  }
-
-  event.preventDefault();
-  event.stopPropagation();
-
-  return true;
-}
-
 interface TransformState {
   didTransform: boolean;
   // Using extra cache (`nodesCurrentlyHighlighting`) since both CodeNode and CodeHighlightNode
@@ -753,18 +310,31 @@ interface TransformState {
   nodesCurrentlyHighlighting: Set<NodeKey>;
 }
 
-export function registerCodeHighlighting(
+/**
+ * @internal
+ * Register only the Prism highlighting transforms and the gutter
+ * mutation listener. No keyboard / indent handlers — those are the
+ * responsibility of
+ * {@link "@lexical/code-core".registerCodeIndentation} /
+ * {@link "@lexical/code-core".CodeIndentExtension}.
+ *
+ * Used by {@link CodePrismExtension}, whose `CodeIndentExtension`
+ * dependency handles the indent side. The legacy
+ * {@link registerCodeHighlighting} wrapper combines this helper with
+ * `registerCodeIndentation` for direct callers that want the original
+ * single-call setup.
+ *
+ * Exported for use by the package's own unit tests; not re-exported
+ * from the package entry point.
+ */
+export function registerHighlightingOnly(
   editor: LexicalEditor,
-  tokenizer?: Tokenizer,
+  tokenizer: Tokenizer,
 ): () => void {
   if (!editor.hasNodes([CodeNode, CodeHighlightNode])) {
     throw new Error(
       'CodeHighlightPlugin: CodeNode or CodeHighlightNode not registered on editor',
     );
-  }
-
-  if (tokenizer == null) {
-    tokenizer = PrismTokenizer;
   }
 
   const registrations = [];
@@ -795,7 +365,6 @@ export function registerCodeHighlighting(
     didTransform: false,
     nodesCurrentlyHighlighting: new Set(),
   };
-  // Add the rest of the registrations
   registrations.push(
     editor.registerNodeTransform(
       CodeNode,
@@ -809,96 +378,26 @@ export function registerCodeHighlighting(
       CodeHighlightNode,
       $textNodeTransform.bind(null, editor, tokenizer, transformState),
     ),
-    editor.registerCommand(
-      KEY_TAB_COMMAND,
-      event => {
-        const command = $handleTab(event.shiftKey);
-        if (command === null) {
-          return false;
-        }
-        event.preventDefault();
-        editor.dispatchCommand(command, undefined);
-        return true;
-      },
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      INSERT_TAB_COMMAND,
-      () => {
-        const selection = $getSelection();
-        if (!$isSelectionInCode(selection)) {
-          return false;
-        }
-        $insertNodes([$createTabNode()]);
-        return true;
-      },
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      INDENT_CONTENT_COMMAND,
-      (payload): boolean => $handleMultilineIndent(INDENT_CONTENT_COMMAND),
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      OUTDENT_CONTENT_COMMAND,
-      (payload): boolean => $handleMultilineIndent(OUTDENT_CONTENT_COMMAND),
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      KEY_ARROW_UP_COMMAND,
-      event => {
-        const selection = $getSelection();
-        if (!$isRangeSelection(selection) || !$isSelectionInCode(selection)) {
-          return false;
-        }
-        const firstNode = $getRoot().getFirstDescendant();
-        const {anchor} = selection;
-        const anchorNode = anchor.getNode();
-        if (
-          firstNode &&
-          anchorNode &&
-          firstNode.getKey() === anchorNode.getKey()
-        ) {
-          return false;
-        }
-        return $handleShiftLines(KEY_ARROW_UP_COMMAND, event);
-      },
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      KEY_ARROW_DOWN_COMMAND,
-      event => {
-        const selection = $getSelection();
-        if (!$isRangeSelection(selection) || !$isSelectionInCode(selection)) {
-          return false;
-        }
-        const lastNode = $getRoot().getLastDescendant();
-        const {anchor} = selection;
-        const anchorNode = anchor.getNode();
-        if (
-          lastNode &&
-          anchorNode &&
-          lastNode.getKey() === anchorNode.getKey()
-        ) {
-          return false;
-        }
-        return $handleShiftLines(KEY_ARROW_DOWN_COMMAND, event);
-      },
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      MOVE_TO_START,
-      event => $handleMoveTo(MOVE_TO_START, event),
-      COMMAND_PRIORITY_LOW,
-    ),
-    editor.registerCommand(
-      MOVE_TO_END,
-      event => $handleMoveTo(MOVE_TO_END, event),
-      COMMAND_PRIORITY_LOW,
-    ),
   );
 
   return mergeRegister(...registrations);
+}
+
+/**
+ * Register the Prism tokenizer-driven highlighting on the editor along
+ * with the indent / Tab / arrow-key keyboard handlers (via
+ * {@link "@lexical/code-core".registerCodeIndentation}). This wrapper
+ * preserves the legacy single-call setup; existing callers do not need
+ * to change.
+ */
+export function registerCodeHighlighting(
+  editor: LexicalEditor,
+  tokenizer?: Tokenizer,
+): () => void {
+  return mergeRegister(
+    registerHighlightingOnly(editor, tokenizer ?? PrismTokenizer),
+    registerCodeIndentation(editor),
+  );
 }
 
 export interface CodePrismConfig {
@@ -917,6 +416,9 @@ export interface CodePrismConfig {
  *
  * {@link CodeExtension} is a dependency, so the required `CodeNode` and
  * `CodeHighlightNode` nodes are registered automatically.
+ * {@link CodeIndentExtension} is also a dependency, so Tab / Shift+Tab
+ * and the related keyboard handlers are activated automatically. Set
+ * `tabSize` on `CodeIndentExtension` to enable space-indent outdent.
  */
 export const CodePrismExtension = defineExtension({
   build: (editor, config) => namedSignals(config),
@@ -924,7 +426,7 @@ export const CodePrismExtension = defineExtension({
     disabled: false,
     tokenizer: PrismTokenizer,
   }),
-  dependencies: [CodeExtension],
+  dependencies: [CodeExtension, CodeIndentExtension],
   name: '@lexical/code-prism',
   register: (editor, config, state) => {
     const stores = state.getOutput();
@@ -932,7 +434,7 @@ export const CodePrismExtension = defineExtension({
       if (stores.disabled.value) {
         return;
       }
-      return registerCodeHighlighting(editor, stores.tokenizer.value);
+      return registerHighlightingOnly(editor, stores.tokenizer.value);
     });
   },
 });
