+++ packages/lexical-code-shiki/src/__tests__/unit/LexicalCodeNodeTabs.test.ts
@@ -14,6 +14,7 @@ import {
   $isCodeHighlightNode,
   $isCodeNode,
 } from '@lexical/code';
+import {registerCodeIndentation} from '@lexical/code-core';
 import {
   isCodeLanguageLoaded,
   loadCodeLanguage,
@@ -42,6 +43,11 @@ import {
 } from 'lexical/src/__tests__/utils';
 import {describe, expect, test} from 'vitest';
 
+import {
+  $runOutdentScenario,
+  OUTDENT_SCENARIOS,
+} from '../../../../lexical-code-core/src/__tests__/outdentTestUtils';
+import {registerHighlightingOnly} from '../../CodeHighlighterShiki';
 import {isCodeThemeLoaded} from '../../FacadeShiki';
 
 describe('LexicalCodeNode tests', () => {
@@ -278,6 +284,43 @@ describe('LexicalCodeNode tests', () => {
           });
         });
       });
+    });
+
+    describe('tabSize (#8410): outdent space-indented lines', () => {
+      test.for(OUTDENT_SCENARIOS)(
+        '$name',
+        async ({
+          rawText,
+          cursorOffset,
+          tabSize,
+          expectedText,
+          expectedCursor,
+        }) => {
+          const {editor} = testEnv;
+          const {text, cursor} = await $runOutdentScenario(
+            editor,
+            async () => {
+              await loadCodeLanguage(ShikiTokenizer.defaultLanguage);
+              await loadCodeTheme(ShikiTokenizer.defaultTheme);
+              registerRichText(editor);
+              registerTabIndentation(editor);
+              // Use the internal highlighting-only helper paired with a
+              // tabSize-aware registerCodeIndentation call. The legacy
+              // wrapper registerCodeHighlighting bundles a no-tabSize
+              // registerCodeIndentation, which would shadow the
+              // tabSize-aware command at the same priority.
+              registerHighlightingOnly(editor, ShikiTokenizer);
+              registerCodeIndentation(editor, tabSize);
+            },
+            rawText,
+            cursorOffset,
+          );
+          expect(text).toBe(expectedText);
+          if (expectedCursor !== undefined) {
+            expect(cursor).toBe(expectedCursor);
+          }
+        },
+      );
     });
   });
 });
